+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.13.8
+
+* Fixed coupon return transactions using audit model
+
+## 1.13.7
+
+* Added auditing for coupon handouts and returns
+* Improved coupon managmenet UI
+
 ## 1.13.6
 
 * Added translations for bank actions